Clean up Header component
