+++ frontend/src/auth-page-components/AuthModal.jsx
@@ -10,7 +10,7 @@ const AuthModal = ({ isOpen, onClose }) => {
       <div className="bg-white rounded-lg p-8 max-w-md w-full relative">
         <button
           onClick={onClose}
-          className="absolute top-4 right-4 text-gray-500 hover:text-gray-700"
+          className="absolute top-4 right-4 text-gray-500 hover:text-gray-700 cursor-pointer"
         >
           <X className="h-6 w-6" />
         </button>
