+++ frontend/src/shared-components/Header.jsx
@@ -1,17 +1,12 @@
-import React, { useState } from "react";
+import React from "react";
 import { Link } from "react-router-dom";
 import SignOutButton from "./SignOutButton";
+import SignInButton from "./SignInButton";
 import { useAuth } from "../context/AuthContext";
-import AuthModal from "../auth-page-components/AuthModal";
 import logo from "/src/assets/finance-companion-logo.png";
 
 const Header = () => {
   const { user } = useAuth();
-  const [showAuthModal, setShowAuthModal] = useState(false);
-
-  const handleSignInClick = () => {
-    setShowAuthModal(true);
-  };
 
   return (
     <header className="flex items-center justify-between px-6 py-4 bg-slate-100 shadow-md">
@@ -28,22 +23,8 @@ const Header = () => {
       </div>
 
       <div className="flex items-center">
-        {user ? (
-          <SignOutButton />
-        ) : (
-          <button
-            onClick={handleSignInClick}
-            className="text-sm font-medium text-royal border border-royal px-4 py-1.5 rounded-md hover:bg-royal hover:text-white transition mr-4"
-          >
-            Sign In
-          </button>
-        )}
+        {user ? <SignOutButton /> : <SignInButton />}
       </div>
-
-      <AuthModal
-        isOpen={showAuthModal}
-        onClose={() => setShowAuthModal(false)}
-      />
     </header>
   );
 };
